feat: push to L instead of using indexes
This makes the logic more flexible for different powers but comes at the
cost of slightly higher opcode cost of push. Given the total cost of the
program, it doesn't really matter that it's slightly more expensive.
Note that we can't have both dynamic size and index assignment because
of https://github.com/algorandfoundation/puya-ts/issues/252